Turn SpecialMainTableColumns into non-static builder
