+++ src/main/java/net/sf/jabref/gui/maintable/MainTableFormat.java
@@ -79,21 +79,22 @@ public class MainTableFormat implements TableFormat<BibEntry> {
         // clear existing column configuration
         tableColumns.clear();
 
+        SpecialMainTableColumnsBuilder builder = new SpecialMainTableColumnsBuilder();
         // Add numbering column to tableColumns
-        tableColumns.add(SpecialMainTableColumns.NUMBER_COL);
+        tableColumns.add(builder.buildNumberColumn());
 
         // Add all file based columns
         if (Globals.prefs.getBoolean(JabRefPreferences.FILE_COLUMN)) {
-            tableColumns.add(SpecialMainTableColumns.FILE_COLUMN);
+            tableColumns.add(builder.buildFileColumn());
         }
 
         if (Globals.prefs.getBoolean(JabRefPreferences.URL_COLUMN)) {
             if (Globals.prefs.getBoolean(JabRefPreferences.PREFER_URL_DOI)) {
-                tableColumns.add(SpecialMainTableColumns
+                tableColumns.add(builder
                         .createIconColumn(JabRefPreferences.URL_COLUMN, MainTableFormat.DOI_FIRST,
                                 new JLabel(IconTheme.JabRefIcon.DOI.getSmallIcon())));
             } else {
-                tableColumns.add(SpecialMainTableColumns
+                tableColumns.add(builder
                         .createIconColumn(JabRefPreferences.URL_COLUMN, MainTableFormat.URL_FIRST,
                                 new JLabel(IconTheme.JabRefIcon.WWW.getSmallIcon())));
             }
@@ -101,7 +102,7 @@ public class MainTableFormat implements TableFormat<BibEntry> {
         }
 
         if (Globals.prefs.getBoolean(JabRefPreferences.ARXIV_COLUMN)) {
-            tableColumns.add(SpecialMainTableColumns
+            tableColumns.add(builder
                     .createIconColumn(JabRefPreferences.ARXIV_COLUMN, MainTableFormat.ARXIV,
                             new JLabel(IconTheme.JabRefIcon.WWW.getSmallIcon())));
         }
@@ -109,7 +110,7 @@ public class MainTableFormat implements TableFormat<BibEntry> {
         if (Globals.prefs.getBoolean(JabRefPreferences.EXTRA_FILE_COLUMNS)) {
             List<String> desiredColumns = Globals.prefs.getStringList(JabRefPreferences.LIST_OF_FILE_COLUMNS);
             for (String desiredColumn : desiredColumns) {
-                tableColumns.add(SpecialMainTableColumns.createFileIconColumn(desiredColumn));
+                tableColumns.add(builder.createFileIconColumn(desiredColumn));
             }
         }
 
@@ -129,22 +130,22 @@ public class MainTableFormat implements TableFormat<BibEntry> {
         // Add the "special" icon columns (e.g., ranking, file, ...) that are enabled in preferences.
         if (Globals.prefs.getBoolean(JabRefPreferences.SPECIALFIELDSENABLED)) {
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_RANKING)) {
-                tableColumns.add(SpecialMainTableColumns.RANKING_COLUMN);
+                tableColumns.add(builder.buildRankingColumn());
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_RELEVANCE)) {
-                tableColumns.add(SpecialMainTableColumns.RELEVANCE_COLUMN);
+                tableColumns.add(builder.buildRelevanceColumn());
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_QUALITY)) {
-                tableColumns.add(SpecialMainTableColumns.QUALITY_COLUMN);
+                tableColumns.add(builder.buildQualityColumn());
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_PRIORITY)) {
-                tableColumns.add(SpecialMainTableColumns.PRIORITY_COLUMN);
+                tableColumns.add(builder.buildPriorityColumn());
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_PRINTED)) {
-                tableColumns.add(SpecialMainTableColumns.PRINTED_COLUMN);
+                tableColumns.add(builder.buildPrintedColumn());
             }
             if (Globals.prefs.getBoolean(JabRefPreferences.SHOWCOLUMN_READ)) {
-                tableColumns.add(SpecialMainTableColumns.READ_STATUS_COLUMN);
+                tableColumns.add(builder.buildReadStatusColumn());
             }
         }
